game-core/movement fix movement generator part1

+++ game-core/src/main/java/com/github/azeroth/game/networking/opcode/OpCode.java
@@ -2,10 +2,10 @@ package com.github.azeroth.game.networking.opcode;
 
 public sealed interface OpCode permits ClientOpCode, ServerOpCode {
 
-    int MAX_OPCODE = 0x3FFF;
-    int NUM_OPCODE_HANDLERS = (MAX_OPCODE + 1);
-    int UNKNOWN_OPCODE = 0xFFFF;
-    int NULL_OPCODE = 0xBADD;
+    short MAX_OPCODE = 0x3FFF;
+    short NUM_OPCODE_HANDLERS = (MAX_OPCODE + 1);
+    short UNKNOWN_OPCODE = (short) 0xFFFF;
+    short NULL_OPCODE = (short) 0xBADD;
 
 
     default boolean isClientToServer() {
